+++ tests/_craft/config/db.php
@@ -2,9 +2,9 @@
 declare(strict_types=1);
 
 return [
-    'dsn' => getenv('DB_DSN'),
-    'user' => getenv('DB_USER'),
-    'password' => getenv('DB_PASSWORD'),
-    'schema' => getenv('DB_SCHEMA'),
-    'tablePrefix' => getenv('DB_TABLE_PREFIX'),
+    'dsn' => 'mysql:host=craft_twig_linter_test_db;port=3306;dbname=craft-twig-linter_testing',
+    'user' => 'homestead',
+    'password' => 'secret',
+    'schema' => '',
+    'tablePrefix' => '',
 ];